+++ asm/src/arg_indirect.c
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   arg_indirect.c                                     :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: fblom <marvin@codam.nl>                      +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2019/12/02 12:02:05 by fblom         #+#    #+#                 */
+/*   Updated: 2019/12/02 12:02:05 by fblom         ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "asm.h"
+
+/*void		save_indirect(char *argument, t_list *item)
+{
+	;
+}
+
+int			check_indirect(char *argument)
+{
+	return (1);
+}
+
+*/
