+++ asm/src/arguments.c
@@ -22,86 +22,6 @@ void		initaliase_values_to_null(t_list *item)
 	(((t_op *)(item->content))->argvalues[2].label) = NULL;
 }
 
-/*void		save_indirect(char *argument, t_list *item)
-{
-	;
-}
-
-int			check_indirect(char *argument)
-{
-	return (1);
-}
-
-*/
-void		save_direct(char *argument, t_list *item, int argno)
-{
-	if (ft_isdigit(argument[0]) || argument[0] == '-')
-		(((t_op *)(item->content))->argvalues[argno]).value = ft_atoi(argument);
-	else
-		(((t_op *)(item->content))->argvalues[argno]).label = ft_strdup(argument + 1);
-}
-
-int			check_direct(char *argument)
-{
-	int		i;
-	char	*digit_string;
-
-	digit_string = ft_strmap(argument, (char (*)(char)) &ft_isdigit);
-	i = 1;
-	if (argument[0] == ':')
-	{
-		while (argument[i])
-		{
-			if (!ft_strchr(LABEL_CHARS, argument[i]))
-				return false;
-			i++;
-		}
-	}
-	else
-	{
-		i = 0;
-		if (argument[i] == '-')
-			i++;
-		while (argument[i] != '\0')
-		{
-			if (digit_string[i] == '\0')
-			{
-				free(digit_string);
-				return (false);
-			}
-			i++;
-		}
-	}
-	free(digit_string);
-	return (true);
-}
-
-
-void		save_register(char *argument, t_list *item, int argno)
-{
-	(((t_op *)(item->content))->argvalues[argno]).value = ft_atoi(argument + 1);
-}
-
-int			check_register(char *argument)
-{
-	int		i;
-	char	*digit_string;
-
-	digit_string = ft_strmap(argument + 1, (char (*)(char)) &ft_isdigit);
-	i = 0;
-	while (argument[i + 1] != '\0')
-	{
-		if (digit_string[i] == '\0')
-		{
-			free(digit_string);
-			return (false);
-		}
-		i++;
-	}
-	free(digit_string);
-	return (true);
-}
-
 void		save_argument_value(char *argument, int type, t_list *item, int argno)
 {
 	char	*trimmed;
